requirements.txt
PyYAML==5.1.2
Click==7.0
Jinja2==2.10.1
